requirements.txt
shapely
svgpathtools
matplotlib
reportlab
rectpack
pillow
webcolors
